backend/requirements.txt
# FastAPI 백엔드 (venv에서 pip install -r requirements.txt)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
openai>=1.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
